# project-implementation/contracts/openapi/content.yaml
...
openapi: 3.1.0
info:
  title: WEMOVE 内容管理 API
  version: 1.1.0
  description: >-
    成员 E CMS；媒体与 PDF 文件参见 files.yaml。所有富文本在服务端用 jsoup 清洗：允许
    p/br/strong/b/em/i/u/s/h2/h3/h4/ul/ol/li/blockquote/a/code/pre/hr，链接仅合法站内路径及 HTTP/HTTPS。
    文章正文另外允许 section/figure/figcaption/img/div；section 仅保留 data-layout=image-left 或 image-right，
    无有效布局的 section 展开为子内容；img 仅保留 src/alt。图片源必须精确匹配
    /assets/products/guides/{balance-stones,rainbow-arch,ring-toss,team-board,forest-kit,skip-rope}-guide.png
    六个内置路径之一，或 /api/v1/media/{小写标准 UUID}/content；其他图片被移除，查询参数、片段及外链均不允许。
    禁止任意 style/class/id、事件、srcset、script、iframe、SVG。FAQ 与站点设置仍不支持正文图片。
    文章保存和发布时从清洗后的正文提取媒体 ID，与显式 mediaIds 去重合并；显式顺序优先，正文按出现顺序追加，
    合计最多 20 张。全部引用须在媒体库存在，草稿引用保护删除，发布后允许公开读取，下架后撤销此文章的公开引用。
    Banner targetUrl 仅站内 / 路径及绝对 HTTP/HTTPS。公开与管理请求均 no-store，写入与媒体引用、审计、幂等同事务。
    商品关联仅接受已发布商品。所有管理路由须 ADMIN 与有效会话；写入须 CSRF 和允许的 Origin。
servers:
- url: "/api/v1"
paths:
  "/admin/articles":
    get:
      operationId: adminListArticles
      summary: 管理员查询Article
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/ArticleView"
                  meta:
                    "$ref": "#/components/schemas/PageMeta"
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
      - name: keyword
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - OFFLINE
      - name: category
        in: query
        required: false
        schema:
          type: string
          maxLength: 80
      - name: productId
        in: query
        required: false
        schema:
          type: string
          format: uuid
    post:
      operationId: createArticle
      summary: 创建草稿Article
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
        '201':
          description: 成功；创建原始请求返回 201，同键同内容重放返回 200。
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ArticleCreate"
  "/admin/articles/{id}":
    get:
      operationId: adminGetArticle
      summary: 管理员读取Article
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateArticle
      summary: 编辑Article；字段作为完整表单提交，已发布内容保存后立即更新公开投影
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ArticleUpdate"
  "/admin/articles/{id}/publish":
    post:
      operationId: publishArticle
      summary: 发布Article
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VersionCommand"
  "/admin/articles/{id}/unpublish":
    post:
      operationId: unpublishArticle
      summary: 下线Article
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VersionCommand"
  "/articles":
    get:
      operationId: publicListArticles
      summary: 仅查询已发布Article
      security: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/ArticleView"
                  meta:
                    "$ref": "#/components/schemas/PageMeta"
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
      - name: keyword
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
      - name: category
        in: query
        required: false
        schema:
          type: string
          maxLength: 80
      - name: productId
        in: query
        required: false
        schema:
          type: string
          format: uuid
  "/articles/{id}":
    get:
      operationId: publicGetArticle
      summary: 读取已发布Article；草稿/下线统一 404
      security: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/ArticleView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
  "/admin/faqs":
    get:
      operationId: adminListFaqs
      summary: 管理员查询Faq
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    type: array
                    items:
                      "$ref": "#/components/schemas/FaqView"
                  meta:
                    "$ref": "#/components/schemas/PageMeta"
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
      - name: keyword
        in: query
        required: false
        schema:
          type: string
          maxLength: 100
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
          - OFFLINE
      - name: category
        in: query
        required: false
        schema:
          type: string
          maxLength: 80
      - name: productId
        in: query
        required: false
        schema:
          type: string
          format: uuid
    post:
      operationId: createFaq
      summary: 创建草稿Faq
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/FaqView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
        '201':
          description: 成功；创建原始请求返回 201，同键同内容重放返回 200。
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/FaqView"
                  meta:
                    type: 'null'
                required:
                - data
      parameters:
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FaqCreate"
  "/admin/faqs/{id}":
    get:
      operationId: adminGetFaq
      summary: 管理员读取Faq
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/FaqView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      operationId: updateFaq
      summary: 编辑Faq；字段作为完整表单提交，已发布内容保存后立即更新公开投影
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/FaqView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/FaqUpdate"
  "/admin/faqs/{id}/publish":
    post:
      operationId: publishFaq
      summary: 发布Faq
      security:
      - session: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/FaqView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: Idempotency-Key
        in: header
        required: true
        schema:
          type: string
          format: uuid
      - name: X-CSRF-Token
        in: header
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VersionCommand"
...
  "/faqs/{id}":
    get:
      operationId: publicGetFaq
      summary: 读取已发布Faq；草稿/下线统一 404
      security: []
      responses:
        '200':
          description: 成功
          headers:
            Cache-Control:
              schema:
                type: string
              description: no-store
            Idempotency-Replayed:
              schema:
                type: boolean
              description: 幂等成功重放时为 true
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  data:
                    "$ref": "#/components/schemas/FaqView"
                  meta:
                    type: 'null'
                required:
                - data
        '401':
          "$ref": "#/components/responses/Error401"
        '403':
          "$ref": "#/components/responses/Error403"
        '404':
          "$ref": "#/components/responses/Error404"
        '409':
          "$ref": "#/components/responses/Error409"
        '422':
          "$ref": "#/components/responses/Error422"
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
...
components:
  responses:
    Error401:
      description: 未登录
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Error403:
      description: 非管理员或 CSRF 校验失败
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Error404:
      description: 资源不存在或未公开
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Error409:
      description: VERSION_CONFLICT / IDEMPOTENCY_CONFLICT / INVALID_STATE_TRANSITION
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
    Error422:
      description: VALIDATION_ERROR，包含字段错误
      content:
        application/problem+json:
          schema:
            "$ref": "#/components/schemas/Problem"
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: WEMOVE_SESSION
  schemas:
    PageMeta:
      type: object
      additionalProperties: false
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
        totalItems:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          format: int64
          minimum: 0
      required:
      - page
      - pageSize
      - totalItems
      - totalPages
...